(maint) update to ember 4.12.1, address issues
This updates to ember 4.12.1 compatibility. This involved
moving from mutable arrays to full array replacement to allow
array tracking to work correctly.

It also involved converting the tab list component into a glimmer
component instead of an ember component.  Some of the argument
signatures changed because of this change, including passing
the model with `@`.

v4.8.0...v4.12.1