(maint) update components to modern implementation
This removes the use of ember run loop functions in favor of
ember-lifeline. It also uses modifiers instead of constructor functions
for the components to register with parents.

It updates the tests to use simple instead of index as the route
as the different pages had coincidental handlers and it was
causing issues.

Tests were also modernized.