(maint) Limit scope and timing of selection
Selection is done by the caller as a part of the callbacks registered
via the tab component and tabs. The timing of this callback depeends
on when the tabs are rendered. This applies a bottleneck to ensure
that the tab selection is honored when specified, by delaying selection
until after the render loop for the current tabs are rendered.

i# Changes to be committed: